test(auth): split oauth finalize api coverage
